requirements.txt
Flask==0.10.1
Flask-Heroku-RQify==0.2
Flask-RQ==0.2
Flask-Script==2.0.5
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.1
beautifulsoup4==4.3.2
click==3.3
gunicorn==19.2.1
itsdangerous==0.24
peewee==2.4.7
psycopg2==2.6
python-dateutil==2.4.0
redis==2.10.3
requests==2.5.1
rq==0.5.0
six==1.9.0
wsgiref==0.1.2
